Refactor to use Clojure watches
This patch replaces the existing polling implementation of SimpleDB with
Clojure's reference watcher functionality. When the backing atom is
swap!ed, this triggers the watcher fn which aquires a lock and flushes
the database. This means that the database will be flushed eagerly on
write rather than somewhat continuously flushed on a real time clock
tick.

While undoubtably slower, this is also saner so deal.